Update 2019-12-05-usr-etc.md
Improved English sentence structure.

The paragraph in the "Atomic Updates" section about admin changes not visible in the update environment is really outdated, so I've rewritten major parts of it.

Reordered parts of the "Goal", "Requirements" and "Solution" sections into other sections for an (imho) more logical structure. Also added a few more sentences about the proposed solution to make clear we want to separate package and admin space.